pyproject.toml
[project]
name = "mcpservertest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.1",
    "mcp[cli]>=1.6.0",
    "pygithub>=2.6.1",
]
